pyproject.toml
[tool.poetry]
name = "twitch-radio"
version = "1.3.0"
description = "Discord bot that plays audio from a Twitch stream"
authors = ["Tomás Farías <tomas@tomasfarias.dev>"]
packages = [
    { include="twitch_radio", from="." },
]

[tool.poetry.scripts]
twitch-radio = "twitch_radio.__main__:main"

[tool.poetry.dependencies]
python = "^3.6"
"discord.py" = "^1.4.1"
streamlink = "^1.5.0"
PyNaCl = "^1.4.0"

[tool.poetry.dev-dependencies]
pytest = "^6.0.1"
flake8 = "^3.8.3"
mypy = "^0.782"

[tool.black]
line-length = 110
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.pytest_cache
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
